@@ -1,13 +1,11 @@
 apply plugin: 'com.android.library'
-apply plugin: 'com.novoda.bintray-release'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion '25.0.0'
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -22,18 +20,13 @@ android {
     }
 }
 
-publish {
-    userOrg = 'qstumn'
-    groupId = 'q.rorbin'
-    artifactId = 'VerticalTabLayout'
-    publishVersion = '1.2.3'
-    desc = 'This is a vertical direction TabLayout'
-    website = 'https://github.com/qstumn/VerticalTabLayout'
-}
-
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    testCompile 'junit:junit:4.12'
-    provided 'com.android.support:support-v4:25.3.0'
-    compile 'q.rorbin:badgeview:1.1.2'
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'androidx.annotation:annotation:1.1.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.viewpager2:viewpager2:1.0.0'
+    implementation 'com.google.android.material:material:1.1.0'
+    testImplementation 'junit:junit:4.13'
+//    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'q.rorbin:badgeview:1.1.3'
 }
