@@ -5,23 +5,14 @@ import android.content.res.TypedArray;
 import android.graphics.Color;
 import android.graphics.drawable.Drawable;
 import android.os.Build;
-import android.support.annotation.Px;
-import android.support.annotation.RequiresApi;
+import androidx.annotation.Px;
+import androidx.annotation.RequiresApi;
 import android.text.TextUtils;
-import android.util.Log;
 import android.view.Gravity;
-import android.view.View;
-import android.widget.Checkable;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import q.rorbin.badgeview.Badge;
-import q.rorbin.badgeview.QBadgeView;
-import q.rorbin.verticaltablayout.util.DisplayUtil;
-
-import static android.R.attr.checked;
-import static android.R.attr.gravity;
 
 /**
  * @author chqiu
